Final_version_v2
