+++ src/components/app/App.tsx
@@ -1,0 +1,37 @@
+import { CSSProperties, useState } from 'react';
+
+import { Article } from '../article/Article';
+import { ArticleParamsForm } from '../article-params-form/ArticleParamsForm';
+import {
+	ArticleStateType,
+	defaultArticleState,
+} from 'src/constants/articleProps';
+
+import styles from './App.module.scss';
+
+export const App = () => {
+	const [articleState, setArticleState] =
+		useState<ArticleStateType>(defaultArticleState);
+	const handleAppChange = (currentFormState: ArticleStateType) => {
+		setArticleState(currentFormState);
+	};
+	return (
+		<main
+			className={styles.main}
+			style={
+				{
+					'--font-family': articleState.fontFamilyOption.value,
+					'--font-size': articleState.fontSizeOption.value,
+					'--font-color': articleState.fontColor.value,
+					'--container-width': articleState.contentWidth.value,
+					'--bg-color': articleState.backgroundColor.value,
+				} as CSSProperties
+			}>
+			<ArticleParamsForm
+				formState={articleState}
+				setFormState={handleAppChange}
+			/>
+			<Article />
+		</main>
+	);
+};
