+++ src/index.tsx
@@ -1,43 +1,11 @@
 import { createRoot } from 'react-dom/client';
-import { StrictMode, CSSProperties, useState } from 'react';
-import clsx from 'clsx';
-
-import { Article } from './components/article/Article';
-import { ArticleParamsForm } from './components/article-params-form/ArticleParamsForm';
-import {
-	ArticleStateType,
-	defaultArticleState,
-} from './constants/articleProps';
+import { StrictMode } from 'react';
+import { App } from './components/app/App';
 
 import './styles/index.scss';
-import styles from './styles/index.module.scss';
 
 const domNode = document.getElementById('root') as HTMLDivElement;
 const root = createRoot(domNode);
-
-const App = () => {
-	const [appState, setAppState] =
-		useState<ArticleStateType>(defaultArticleState);
-	const handleAppChange = (currentFormState: ArticleStateType) => {
-		setAppState(currentFormState);
-	};
-	return (
-		<main
-			className={clsx(styles.main)}
-			style={
-				{
-					'--font-family': appState.fontFamilyOption.value,
-					'--font-size': appState.fontSizeOption.value,
-					'--font-color': appState.fontColor.value,
-					'--container-width': appState.contentWidth.value,
-					'--bg-color': appState.backgroundColor.value,
-				} as CSSProperties
-			}>
-			<ArticleParamsForm formState={appState} setFormState={handleAppChange} />
-			<Article />
-		</main>
-	);
-};
 
 root.render(
 	<StrictMode>
